Rename Concurrent to ConcurrentIterator
