+++ composer.json
@@ -35,7 +35,7 @@
         },
         "files": [
             "src/functions.php",
-            "src/Concurrent/functions.php"
+            "src/ConcurrentIterator/functions.php"
         ]
     },
     "autoload-dev": {
